I think this is code completion, maybe just a couple small things to add.

+++ mixedvr-manager.bat
@@ -47,7 +47,7 @@ if "%ERRORLEVEL%"=="0" (set steamvrStatus=running) else (set steamvrStatus=quit)
 :: handle the first loop case; we'll only detect changes to the running status since the script started, not on first run
 if "%steamvrLastKnownStatus%" == "" (
 	set steamvrLastKnownStatus=%steamvrStatus%
-	echo Setting inits, at time of script start, SteamVR was %steamvrStatus%
+	echo Initializing: at time of MixedVR-Manager start, SteamVR was: %steamvrStatus%
 	echo MixedVR-Manager is now waiting for events...
 	goto whileTrueLoop
 )
@@ -61,7 +61,7 @@ if "%steamvrStatus%" == "%steamvrLastKnownStatus%" (
 :: if we got to this point, then that means that SteamVR's status has just changed
 :: so we call our stateChanged function, which effectively passes the the new state
 :: in as "steamvrStatus"
-echo Change detected in SteamVR running status. SteamVR is now %steamvrStatus%
+echo MixedVR-Manager detected that SteamVR is now: %steamvrStatus%
 goto stateChanged
 
 
@@ -77,12 +77,12 @@ goto stateChanged
 
 :: toggle lighthouse state
 if "%steamvrStatus%" == "running" (set desiredLighthouseState=on) else (set desiredLighthouseState=off)
-echo Turning lighthouses %desiredLighthouseState%...
+echo MixedVR-Manager is turning lighthouses %desiredLighthouseState%...
 lighthouse-v2-manager.exe %desiredLighthouseState% %lighthouseMACAddressList%
 
 :: toggle state of the USB that the headset is plugged into
 if "%steamvrStatus%" == "running" (set desiredHMDUSBAction=enable) else (set desiredHMDUSBAction=disable)
-echo Changing state of USB device (the HMD) to /%desiredHMDUSBAction%...
+echo MixedVR-Manager is changing state of USB device (the HMD) to /%desiredHMDUSBAction%...
 USBDeview.exe /RunAsAdmin /%desiredHMDUSBAction% "HoloLens Sensors"
 
 :: if we're switching to the running state, then we also need to restart SteamVR now that 
@@ -90,37 +90,39 @@ USBDeview.exe /RunAsAdmin /%desiredHMDUSBAction% "HoloLens Sensors"
 :: note, we need to kill room setup if it's running because otherwise it may still be open when we 
 :: start relying on it being quit when blocking on the next loop
 if "%steamvrStatus%" == "running" (
-	echo Killing SteamVR and restarting it so it can detect the now powered on lighthouses and HMD.
+	echo MixedVR-Manager is restarting SteamVR so it SteamVR can detect the now powered on lighthouses and HMD.
 	taskkill /f /im "steamvr_room_setup.exe"
 	taskkill /f /im "vrmonitor.exe"
 	taskkill /f /im "vrserver.exe"
+	:: TODO: should kill OpenVR space calibrator here too
 	timeout 1 >NUL
 	start steam://launch/250820/VR
+
+	:: wait until SteamVR Room Setup starts, then kill it. if it doesn't start after 
+	:: maxWaitTimeForRoomSetup seconds, assume it's never going to start, and just 
+	:: continue with the script's execution. 
+	:: note: this "delayed expansion" business is a real treat; apparently variables are 
+	:: evaluated before execution time unless you do this and then use ! instead of %. Craziness.
+	: roomSetupLoop
+	echo Waiting for SteamVR to start back up and to close Room Setup...
+	setlocal EnableDelayedExpansion
+	for /L %%i in (1,1,%maxWaitTimeForRoomSetup%) do (
+		tasklist /FI "IMAGENAME eq steamvr_room_setup.exe" 2>NUL | find /I /N "steamvr_room_setup.exe">NUL
+		if "!ERRORLEVEL!"=="0" (set roomSetupStatus=running) else (set roomSetupStatus=quit)
+		if "!roomSetupStatus!" == "running" (
+			taskkill /f /im "steamvr_room_setup.exe" 
+			goto break
+		) else (
+			timeout 1 >NUL
+		)
+	)
+	:break
+
 ) else (
 	:: this means they just shut down SteamVR, which means we'll also want to clean up 
 	:: any other applications that won't be shut down automatically, like WMR
 	taskkill /f /im "MixedRealityPortal.exe"
 )
-
-:: wait until SteamVR Room Setup starts, then kill it. if it doesn't start after 
-:: maxWaitTimeForRoomSetup seconds, assume it's never going to start, and just 
-:: continue with the script's execution. 
-:: note: this "delayed expansion" business really got me; apparently variables are 
-:: evaluated before execution time unless you do this and then use ! instead of %. Craziness.
-: roomSetupLoop
-echo Waiting for SteamVR to start back up and to close Room Setup...
-setlocal EnableDelayedExpansion
-for /L %%i in (1,1,%maxWaitTimeForRoomSetup%) do (
-	tasklist /FI "IMAGENAME eq steamvr_room_setup.exe" 2>NUL | find /I /N "steamvr_room_setup.exe">NUL
-	if "!ERRORLEVEL!"=="0" (set roomSetupStatus=running) else (set roomSetupStatus=quit)
-	if "!roomSetupStatus!" == "running" (
-		taskkill /f /im "steamvr_room_setup.exe" 
-		goto break
-	) else (
-		timeout 1 >NUL
-	)
-)
-:break
 
 :: mark the new last known state
 set steamvrLastKnownStatus=%steamvrStatus%
